requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.12.0
ollama==0.1.5
Werkzeug==2.3.7
gunicorn==21.2.0
markdown==3.5.1
markupsafe==2.1.3 